+++ CHANGELOG.md
@@ -10,6 +10,7 @@ project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - `ProgressBar` control for tracking the completion of a task
 - `RadioButtons` control for groups of radio buttons
+- `Combobox::selected()` method to retrieve the currently selected index of the combobox
 
 ### Changed